refactor: 기물 배치 전략 인터페이스 도입
