+++ src/main/java/janggi/board/strategy/PlaceStrategy.java
@@ -1,0 +1,11 @@
+package janggi.board.strategy;
+
+import janggi.board.Position;
+import janggi.piece.Piece;
+
+import java.util.Map;
+
+public interface PlaceStrategy {
+
+    Map<Position, Piece> initialize();
+}
