+++ src/test/java/janggi/piece/ElephantTest.java
@@ -26,7 +26,7 @@ class ElephantTest {
         Elephant piece = new Elephant(Team.GREEN);
 
         initialBoard.put(start, piece);
-        Board board = new Board(initialBoard);
+        Board board = new Board(() -> initialBoard);
 
         // when
         board.movePiece(start, goal, Team.GREEN);
@@ -47,7 +47,7 @@ class ElephantTest {
 
         initialBoard.put(start, piece);
         initialBoard.put(createPosition(4,5), new Soldier(Team.GREEN));
-        Board board = new Board(initialBoard);
+        Board board = new Board(() -> initialBoard);
 
         // then
         assertThatThrownBy(() -> board.movePiece(start, goal, Team.GREEN))
